+++ retriever/try_install_all.py
@@ -13,7 +13,8 @@ import os
 import sys
 from imp import reload
 from retriever.lib.tools import choose_engine
-from retriever import MODULE_LIST, ENGINE_LIST, SCRIPT_LIST
+from retriever.lib.scripts import MODULE_LIST, SCRIPT_LIST
+from retriever import ENGINE_LIST
 
 reload(sys)
 if hasattr(sys, 'setdefaultencoding'):
